+++ workbench/hono/nitro.config.ts
@@ -1,9 +1,14 @@
 import { defineConfig } from 'nitro';
 
+const plugins =
+  process.env.WORKFLOW_TARGET_WORLD === '@workflow/world-postgres'
+    ? ['plugins/start-pg-world.ts']
+    : [];
+
 export default defineConfig({
   modules: ['workflow/nitro'],
   routes: {
     '/**': './src/index.ts',
   },
-  plugins: ['plugins/start-pg-world.ts'],
+  plugins,
 });